Fix portal ip alerts
